// App_Scholar/frontend/src/schemas/disciplina.schema.ts
import { z } from 'zod';
export const disciplinaSchema = z.object({
  disciplinaNome: z.string().min(1, "Nome da disciplina é obrigatório"),
  disciplinaCargaHoraria: z.coerce.number().min(1, "Carga horária é obrigatória"),
  disciplinaSemestre: z.coerce.number().min(1, "Semestre é obrigatório"),
  professorId: z.string().min(1, "Professor é obrigatório"),
  cursoId: z.string().min(1, "Curso é obrigatório"),
});

export type DisciplinaFormData = z.infer<typeof disciplinaSchema>;